Move CNI code into plugin package.
