+++ cni/pkg/plugin/iptables.go
@@ -1,4 +1,4 @@
-// Copyright 2019 Istio Authors
+// Copyright Istio Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 
 // This is a sample chained plugin that supports multiple CNI versions. It
 // parses prevResult according to the cniVersion
-package main
+package plugin
 
 import (
 	"fmt"
