+++ cni/pkg/plugin/redirect.go
@@ -1,4 +1,4 @@
-// Copyright 2018 Istio authors
+// Copyright Istio Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 // Defines the redirect object and operations.
-package main
+package plugin
 
 import (
 	"fmt"
